+++ src/ws/stream.ts
@@ -1,27 +1,60 @@
-import { ClientError } from "../errors.js";
+import { ClientError, ClosedError, InternalError } from "../errors.js";
+import { Queue } from "../queue.js";
 import type { SqlOwner, ProtoSql } from "../sql.js";
 import { Stream } from "../stream.js";
 
 import type { WsClient } from "./client.js";
-import type { CursorState } from "./cursor.js";
 import { WsCursor } from "./cursor.js";
 import type * as proto from "./proto.js";
 
-export interface StreamState {
-    streamId: number;
-    closed: Error | undefined;
-    cursorState: CursorState | undefined;
+type QueueEntry = RequestEntry | CursorEntry;
+
+type RequestEntry = {
+    type: "request",
+    request: proto.Request,
+    responseCallback: (_: proto.Response) => void;
+    errorCallback: (_: Error) => void;
+}
+
+type CursorEntry = {
+    type: "cursor",
+    batch: proto.Batch,
+    cursorCallback: (_: WsCursor) => void,
+    errorCallback: (_: Error) => void,
 }
 
 export class WsStream extends Stream {
     #client: WsClient;
-    #state: StreamState;
+    #streamId: number;
+
+    #queue: Queue<QueueEntry>;
+    #cursor: WsCursor | undefined;
+    #closing: boolean;
+    #closed: Error | undefined;
 
     /** @private */
-    constructor(client: WsClient, state: StreamState) {
+    static open(client: WsClient): WsStream {
+        const streamId = client._streamIdAlloc.alloc();
+        const stream = new WsStream(client, streamId);
+
+        const responseCallback = () => undefined;
+        const errorCallback = (e: Error) => stream.#setClosed(e);
+
+        const request: proto.OpenStreamReq = {type: "open_stream", streamId};
+        client._sendRequest(request, {responseCallback, errorCallback});
+        return stream;
+    }
+
+    /** @private */
+    constructor(client: WsClient, streamId: number) {
         super(client.intMode);
         this.#client = client;
-        this.#state = state;
+        this.#streamId = streamId;
+
+        this.#queue = new Queue();
+        this.#cursor = undefined;
+        this.#closing = false;
+        this.#closed = undefined;
     }
 
     /** Get the {@link WsClient} object that this stream belongs to. */
@@ -38,7 +71,7 @@ export class WsStream extends Stream {
     override _execute(stmt: proto.Stmt): Promise<proto.StmtResult> {
         return this.#sendStreamRequest({
             type: "execute",
-            streamId: this.#state.streamId,
+            streamId: this.#streamId,
             stmt,
         }).then((response) => {
             return (response as proto.ExecuteResp).result;
@@ -49,7 +82,7 @@ export class WsStream extends Stream {
     override _batch(batch: proto.Batch): Promise<proto.BatchResult> {
         return this.#sendStreamRequest({
             type: "batch",
-            streamId: this.#state.streamId,
+            streamId: this.#streamId,
             batch,
         }).then((response) => {
             return (response as proto.BatchResp).result;
@@ -61,7 +94,7 @@ export class WsStream extends Stream {
         this.#client._ensureVersion(2, "describe()");
         return this.#sendStreamRequest({
             type: "describe",
-            streamId: this.#state.streamId,
+            streamId: this.#streamId,
             sql: protoSql.sql,
             sqlId: protoSql.sqlId,
         }).then((response) => {
@@ -74,18 +107,12 @@ export class WsStream extends Stream {
         this.#client._ensureVersion(2, "sequence()");
         return this.#sendStreamRequest({
             type: "sequence",
-            streamId: this.#state.streamId,
+            streamId: this.#streamId,
             sql: protoSql.sql,
             sqlId: protoSql.sqlId,
         }).then((_response) => {
             return undefined;
         });
-    }
-
-    /** @private */
-    override async _openCursor(batch: proto.Batch): Promise<WsCursor> {
-        const cursorState = this.#client._openCursor(this.#state, batch);
-        return new WsCursor(this.#client, this.#state, cursorState);
     }
 
     /** Check whether the SQL connection underlying this stream is in autocommit state (i.e., outside of an
@@ -95,7 +122,7 @@ export class WsStream extends Stream {
         this.#client._ensureVersion(3, "getAutocommit()");
         return this.#sendStreamRequest({
             type: "get_autocommit",
-            streamId: this.#state.streamId,
+            streamId: this.#streamId,
         }).then((response) => {
             return (response as proto.GetAutocommitResp).isAutocommit;
         });
@@ -103,17 +130,126 @@ export class WsStream extends Stream {
 
     #sendStreamRequest(request: proto.Request): Promise<proto.Response> {
         return new Promise((responseCallback, errorCallback) => {
-            this.#client._sendStreamRequest(this.#state, request, {responseCallback, errorCallback});
+            this.#pushToQueue({type: "request", request, responseCallback, errorCallback});
         });
     }
 
-    /** Close the stream. */
-    override close(): void {
-        this.#client._closeStream(this.#state, new ClientError("Stream was manually closed"));
+    /** @private */
+    override _openCursor(batch: proto.Batch): Promise<WsCursor> {
+        this.#client._ensureVersion(3, "cursor");
+        return new Promise((cursorCallback, errorCallback) => {
+            this.#pushToQueue({type: "cursor", batch, cursorCallback, errorCallback});
+        });
     }
 
-    /** True if the stream is closed. */
+    /** @private */
+    _sendCursorRequest(cursor: WsCursor, request: proto.Request): Promise<proto.Response> {
+        if (cursor !== this.#cursor) {
+            throw new InternalError("Cursor not associated with the stream attempted to execute a request");
+        }
+        return new Promise((responseCallback, errorCallback) => {
+            if (this.#closed !== undefined) {
+                errorCallback(new ClosedError("Stream is closed", this.#closed));
+            } else {
+                this.#client._sendRequest(request, {responseCallback, errorCallback});
+            }
+        });
+    }
+
+    /** @private */
+    _cursorClosed(cursor: WsCursor): void {
+        if (cursor !== this.#cursor) {
+            throw new InternalError("Cursor was closed, but it was not associated with the stream");
+        }
+        this.#cursor = undefined;
+        this.#flushQueue();
+    }
+
+    #pushToQueue(entry: QueueEntry): void {
+        if (this.#closed !== undefined) {
+            entry.errorCallback(new ClosedError("Stream is closed", this.#closed));
+        } else if (this.#closing) {
+            entry.errorCallback(new ClosedError("Stream is closing", undefined));
+        } else {
+            this.#queue.push(entry);
+            this.#flushQueue();
+        }
+    }
+
+    #flushQueue(): void {
+        for (;;) {
+            const entry = this.#queue.first();
+            if (entry === undefined && this.#cursor === undefined && this.#closing) {
+                this.#setClosed(new ClientError("Stream was gracefully closed"));
+                break;
+            } else if (entry?.type === "request" && this.#cursor === undefined) {
+                const {request, responseCallback, errorCallback} = entry;
+                this.#queue.shift();
+
+                this.#client._sendRequest(request, {responseCallback, errorCallback});
+            } else if (entry?.type === "cursor" && this.#cursor === undefined) {
+                const {batch, cursorCallback} = entry;
+                this.#queue.shift();
+
+                const cursorId = this.#client._cursorIdAlloc.alloc();
+                const cursor = new WsCursor(this.#client, this, cursorId);
+
+                const request: proto.OpenCursorReq = {
+                    type: "open_cursor",
+                    streamId: this.#streamId,
+                    cursorId,
+                    batch,
+                };
+                const responseCallback = () => undefined;
+                const errorCallback = (e: Error) => cursor._setClosed(e);
+                this.#client._sendRequest(request, {responseCallback, errorCallback});
+
+                this.#cursor = cursor;
+                cursorCallback(cursor);
+            } else {
+                break;
+            }
+        }
+    }
+
+    #setClosed(error: Error): void {
+        if (this.#closed !== undefined) {
+            return;
+        }
+        this.#closed = error;
+
+        if (this.#cursor !== undefined) {
+            this.#cursor._setClosed(error);
+        }
+
+        for (;;) {
+            const entry = this.#queue.shift();
+            if (entry !== undefined) {
+                entry.errorCallback(error);
+            } else {
+                break;
+            }
+        }
+
+        const request: proto.CloseStreamReq = {type: "close_stream", streamId: this.#streamId};
+        const responseCallback = () => this.#client._streamIdAlloc.free(this.#streamId);
+        const errorCallback = () => undefined;
+        this.#client._sendRequest(request, {responseCallback, errorCallback});
+    }
+
+    /** Immediately close the stream. */
+    override close(): void {
+        this.#setClosed(new ClientError("Stream was manually closed"));
+    }
+
+    /** Gracefully close the stream. */
+    override closeGracefully(): void {
+        this.#closing = true;
+        this.#flushQueue();
+    }
+
+    /** True if the stream is closed or closing. */
     override get closed(): boolean {
-        return this.#state.closed !== undefined;
+        return this.#closed !== undefined || this.#closing;
     }
 }